Made Rust version compile
Changed rust version to include a main function, and prefixed number with an underscore to achieve a compile without warning. Sadly, this adds 1 byte to the length of the is_prime function. This could be prevented by naming the variable _ instead of _number, but this is less consistent across languages.