Clean README.md
I added some `<br />`s after every question in the FAQ section so the questions and answeres are no longer separated by a big gap. I then made the "Guidelines" subtitle a bit bigger, so as not to be confused with a regular piece of text. I also removed some random whitespace.